Changed double to float - should be faster

+++ core/ImgMatch.h
@@ -33,9 +33,9 @@ class ImgMatch
 
     /* Returns match value in the range [0.00, 1.00] */
 #if 0  // Takes the images themselves as parameters
-    virtual double Compare( const Image& img1, const Image& img2 ) = 0;
+    virtual float Compare( const Image& img1, const Image& img2 ) = 0;
 #else  // Takes image locations (paths)
-    virtual double Compare( const std::string& img1path, const std::string& img2path ) = 0;
+    virtual float Compare( const std::string& img1path, const std::string& img2path ) = 0;
 #endif // 0
 
   private:
